Move the ability to import CSVs to the base TestCase
